Change how the master EC2 instance for EMR is detected
Querying by public DNS is impossible if the EMR is in the private
subnet. Instead, the masters would have the master_security_group. Given
that we only have one for now, this should be sufficient enough. For
multiple masters, we will need to revisit this for a @TODO Find first
master AWS instance for EMR.